acabo día 08/04

+++ src/main/java/com/tfg/dao/CampoRepository.java
@@ -10,7 +10,7 @@ import com.tfg.model.Campo;
  * @param <MetadatoRepository>
  *
  */
-public interface CampoRepository extends CrudRepository<Campo, Integer>{
+public interface CampoRepository extends CrudRepository<Campo, Long>{
 	
 	Campo findByCodigo(String nombre);
