+++ src/main/java/com/tfg/dao/GrupoRepository.java
@@ -9,7 +9,7 @@ import com.tfg.model.Grupo;
  * @author gcollada
  *
  */
-public interface GrupoRepository extends CrudRepository<Grupo, Integer>{
+public interface GrupoRepository extends CrudRepository<Grupo, Long>{
 	
 	Grupo findByCodigo(String codigo);
